added apyCalc utility
